# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Content and Project Management API",
    "description": "API for managing content, projects, categories, subcategories, areas and statuses, with history and access control.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "creator"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "subcategories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Subcategory" }
          }
        }
      },
      "Subcategory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "categoryId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "category": { "$ref": "#/components/schemas/Category" }
        }
      },
      "Content": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string", "maxLength": 200 },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["draft", "published", "archived"], "default": "draft" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "authorId": { "type": "integer" },
          "lastUpdatedById": { "type": "integer", "nullable": true },
          "categoryId": { "type": "integer", "nullable": true },
          "subcategoryId": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "author": { "$ref": "#/components/schemas/User" },
          "lastUpdatedBy": { "$ref": "#/components/schemas/User" },
          "category": { "$ref": "#/components/schemas/Category" },
          "subcategory": { "$ref": "#/components/schemas/Subcategory" }
        }
      },
      "ContentHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "action": { "type": "string", "enum": ["created", "updated", "status_changed", "deleted"] },
          "changes": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "previousData": { "type": "string", "nullable": true },
          "newData": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" },
          "content": { "$ref": "#/components/schemas/Content" }
        }
      },
...
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "CreateSubcategoryDto": {
        "type": "object",
        "required": ["name", "categoryId"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "categoryId": { "type": "integer" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateSubcategoryDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "categoryId": { "type": "integer" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
...
      "RegisterDto": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "role": { "type": "string", "enum": ["admin", "creator"], "default": "creator" }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "description": "HTML" },
          "target": { "type": "string", "description": "HTML" },
          "conclusion": { "type": "string", "description": "HTML", "nullable": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "statusId": { "type": "integer" },
          "areaId": { "type": "integer" },
          "authorId": { "type": "integer" },
          "lastUpdatedById": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "author": { "$ref": "#/components/schemas/User" },
          "lastUpdatedBy": { "$ref": "#/components/schemas/User" },
          "status": { "$ref": "#/components/schemas/ProjectStatus" },
          "area": { "$ref": "#/components/schemas/ProjectArea" }
        }
      },
...
      "ProjectStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ProjectArea": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ProjectHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "action": { "type": "string", "enum": ["created", "updated", "status_changed", "area_changed", "conclusion_added", "deleted"] },
          "changes": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "previousData": { "type": "string", "nullable": true },
          "newData": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" },
          "project": { "$ref": "#/components/schemas/Project" }
        }
      }
    }
  },
  "paths": {
...
    "/content/history/user/{userId}": {
      "get": {
        "tags": ["Contenido"],
        "summary": "Obtener historial de un usuario (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ContentHistory" }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/public": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener categorías activas (público)",
        "parameters": [
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "name", "createdAt", "updatedAt"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías activas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "tags": ["Categorías"],
        "summary": "Obtener categoría por ID (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Categorías"],
        "summary": "Actualizar categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCategoryDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Categorías"],
        "summary": "Eliminar categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría eliminada exitosamente"
          }
        }
      }
    },
    "/categories/{id}/toggle": {
      "patch": {
        "tags": ["Categorías"],
        "summary": "Cambiar estado de categoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de categoría cambiado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
    "/subcategories/public": {
      "get": {
        "tags": ["Subcategorías"],
        "summary": "Obtener subcategorías activas (público)",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "Filtrar por categoría",
            "schema": { "type": "integer" }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "name", "createdAt", "updatedAt"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de subcategorías activas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Subcategory" }
                }
              }
            }
          }
        }
      }
    },
    "/subcategories": {
      "get": {
        "tags": ["Subcategorías"],
        "summary": "Obtener todas las subcategorías (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "description": "Filtrar por estado activo",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "Filtrar por categoría",
            "schema": { "type": "integer" }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar",
            "schema": {
              "type": "string",
              "enum": ["priority", "name", "createdAt", "updatedAt"],
              "default": "priority"
            }
          },
          {
            "name": "orderDirection",
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de subcategorías",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Subcategory" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Subcategorías"],
        "summary": "Crear nueva subcategoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSubcategoryDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subcategoría creada exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subcategory" }
              }
            }
          }
        }
      }
    },
    "/subcategories/{id}": {
      "get": {
        "tags": ["Subcategorías"],
        "summary": "Obtener subcategoría por ID (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcategoría encontrada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subcategory" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Subcategorías"],
        "summary": "Actualizar subcategoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSubcategoryDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subcategoría actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subcategory" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Subcategorías"],
        "summary": "Eliminar subcategoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Subcategoría eliminada exitosamente"
          }
        }
      }
    },
    "/subcategories/{id}/toggle": {
      "patch": {
        "tags": ["Subcategorías"],
        "summary": "Cambiar estado de subcategoría (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de subcategoría cambiado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Subcategory" }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener todos los usuarios (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear nuevo usuario (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener usuario por ID (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Usuarios"],
        "summary": "Actualizar usuario (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar usuario (solo ADMIN)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado exitosamente"
          }
        }
      }
    },
...
    "/projects/history/user/{userId}": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Historial de proyectos de un usuario (solo admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Historial del usuario", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectHistory" } } } } }
        }
      }
    },
...
  }
} 